slicer_getbuildinfo: Display number of updated records
Since when using the `insert or replace into` statement below, all records
are effectively replaced independently of their value, the `cursor.rowcount`
property does not allow to know the number of records effectively updated.

To address this, we compute the checksum of all records before and after
and use that to infer the number of records added and updated.